Xamarin prep

+++ src/Domain/Entities/BatchSample.cs
@@ -7,7 +7,7 @@ namespace brewjournal.Domain.Entities
         public int Id { get; set; }
         public int BatchId { get; set; }
         public DateTime SampleDate { get; set; }
-        public long Gravity { get; set; }
+        public float Gravity { get; set; }
         public long Temperature { get; set; }
     }
 }
